Update UI for user registration and pricing plans
Added a click handler to initiate Google Analytics event tracking upon user registration. Updated the "sign up" buttons' click actions to scroll to the pricing section. Enabled the "Packages" component in the main page and made changes to the pricing card component, allowing for the disabling of buttons and adding an optional click handler. Changed the "header" component to link the signup button to the pricing section and update the menu items accordingly. This update enhances user onboarding and provides valuable insights through event tracking.