Update CTA Button Texts and Add Sign-in Menu
Modified the call-to-action (CTA) button texts in various sections for clearer user guidance and added a new sign-in menu in the header. In 'package.js', the 'Start for Free' button text was changed to 'Sign Up' to align with its functionality. In 'banner.js' and 'header.js', 'Sign Up' was replaced with 'Start for free' to emphasize the accessibility of the service. A 'Sign in' menu has been implemented to 'header.data.js' for easier user access. Also, the sitemap.xml was updated as a general maintenance task.