Update chatbot library source and sitemap last modification date
Chatbot script source has been updated to point to a custom location. This was necessary to control the specific version of the script loaded in our web app. In addition, the sitemap.xml file's last modification date has been updated to reflect the latest site changes, improving the accuracy of search engine indexing.