Remove obsolete Google Analytics
Deleted the Google Analytics logic as it was no longer required. This includes disposal of the 'analytics' directory and all related imports and invocations from '_app.js'. The 'yarn.lock' file was updated to remove corresponding redundant dependencies. This action was necessary for better alignment with GDPR compliance as well as reducing back-end clutter and optimizing app performance.